refactor: 에러바운더리 파일명 변경
컴포넌트명에 맞게 변경
